add new src

+++ src/main/java/org/example/BmiService.java
@@ -1,7 +1,7 @@
 package org.example;
 
 public class BmiService {
-    public int calculate(double heightInMeters, int weightInKg) {
+    public int calculate(double heightInMeters, double weightInKg) {
         double bmi = weightInKg / heightInMeters / heightInMeters;
         return (int) bmi;
     }
